Support 'Almost' translated languages
This application is used in various countries. And in Crowdin, some languages are already 100% translated or almost translated. But Glucosio is still support only English. These features are enhancement of it. And also, you can find Crowdin's API for translation status. But there are too many non-translated languages and it was useless to get it's all data. The API needs Auth key so until it is cleared, just maintaining available strings manually would be better. I already made code, but I will just keep it.

- Add available_languages.xml of 'almost' translated languages
- Add updateLocale code on MainPresenter when there is already user profile in DB
- Swap two codes order of calling MainPresenter for update MainActivities in start
- Add localeHelper variable for MainActivity
- Changed getLocalesWithTranslation to available languages
- Fixed bug in build.gradle. It should contain one more backslash